# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TreasureTots Creations e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  role: string;
  createdAt?: string;
}

export interface RegisterInput {
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  password: string;
}

export interface LoginInput {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface ProfileUpdate {
  name?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  price?: number | null;
  stock?: number;
  coverImage: string;
  images: string[];
  category: string;
  /** @nullable */
  subcategory?: string | null;
  slug: string;
  isBuyable: boolean;
  isActive: boolean;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  perPage: number;
}

export interface ProductInput {
  name: string;
  description?: string;
  /** @nullable */
  price?: number | null;
  stock?: number;
  coverImage: string;
  images: string[];
  category: string;
  /** @nullable */
  subcategory?: string | null;
  slug: string;
  isBuyable?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  /** @nullable */
  price?: number | null;
  stock?: number;
  coverImage?: string;
  images?: string[];
  category?: string;
  /** @nullable */
  subcategory?: string | null;
  isBuyable?: boolean;
  isActive?: boolean;
}

export interface CartItem {
  productId: number;
  product: Product;
  quantity: number;
  /** @nullable */
  childName?: string | null;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  quantity?: number;
  /** @nullable */
  childName?: string | null;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface Address {
  id: number;
  fullName: string;
  phone: string;
  houseNo: string;
  street: string;
  city: string;
  state: string;
  pincode: string;
  isDefault?: boolean;
}

export interface AddressInput {
  fullName: string;
  phone: string;
  houseNo: string;
  street: string;
  city: string;
  state: string;
  pincode: string;
  isDefault?: boolean;
}

export interface WishlistInput {
  productId: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  product?: Product;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  totalAmount: number;
  paymentStatus: string;
  orderStatus: string;
  /** @nullable */
  childName?: string | null;
  shippingAddress?: Address;
  /** @nullable */
  razorpayOrderId?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export type OrderInitInputItemsItem = {
  productId: number;
  quantity: number;
};

export interface OrderInitInput {
  items: OrderInitInputItemsItem[];
  /** @nullable */
  childName?: string | null;
  address: AddressInput;
}

export interface OrderInitResponse {
  orderId: number;
  razorpayOrderId: string;
  amount: number;
  key: string;
}

export interface PaymentVerifyInput {
  razorpayPaymentId: string;
  razorpayOrderId: string;
  razorpaySignature: string;
  orderId: number;
}

export interface OrderStatusUpdate {
  orderStatus?: string;
  paymentStatus?: string;
}

export type AnalyticsMonthlyRevenueItem = {
  month?: string;
  revenue?: number;
};

export interface Analytics {
  totalRevenue: number;
  totalOrders: number;
  activeUsers: number;
  /** @nullable */
  topProduct?: string | null;
  monthlyRevenue?: AnalyticsMonthlyRevenueItem[];
}

export type ListProductsParams = {
category?: string;
subcategory?: string;
page?: number;
per_page?: number;
};

export type SearchProductsParams = {
q: string;
};

export type AdminListOrdersParams = {
payment_status?: string;
order_status?: string;
};