# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TreasureTots Creations e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      summary: Admin — list all products
      responses:
        "200":
          description: All products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Admin — create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /v1/admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Admin — update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Admin — soft-delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /v1/admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      summary: Admin — list all orders
      parameters:
        - name: payment_status
          in: query
          schema:
            type: string
        - name: order_status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /v1/admin/orders/{id}/status:
    patch:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Admin — update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
      required: [id, name, role]
    RegisterInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        password:
          type: string
      required: [name, password]
    LoginInput:
      type: object
      properties:
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        password:
          type: string
      required: [password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        stock:
          type: integer
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        slug:
          type: string
        isBuyable:
          type: boolean
        isActive:
          type: boolean
      required: [id, name, coverImage, images, category, slug, isBuyable, isActive]
...
    ProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
        stock:
          type: integer
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        slug:
          type: string
        isBuyable:
          type: boolean
      required: [name, coverImage, images, category, slug]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: ["number", "null"]
        stock:
          type: integer
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        isBuyable:
          type: boolean
        isActive:
          type: boolean
...
    Address:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        houseNo:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
        isDefault:
          type: boolean
      required: [id, fullName, phone, houseNo, street, city, state, pincode]
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price:
          type: number
      required: [id, productId, quantity, price]
    Order:
      type: object
      properties:
        id:
          type: integer
        totalAmount:
          type: number
        paymentStatus:
          type: string
        orderStatus:
          type: string
        childName:
          type: ["string", "null"]
        shippingAddress:
          $ref: "#/components/schemas/Address"
        razorpayOrderId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
      required: [id, totalAmount, paymentStatus, orderStatus, items, createdAt]
...
    OrderStatusUpdate:
      type: object
      properties:
        orderStatus:
          type: string
        paymentStatus:
          type: string
...