# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TreasureTots Creations e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from './address';
import type { OrderItem } from './orderItem';

export interface Order {
  id: number;
  totalAmount: number;
  paymentStatus: string;
  orderStatus: string;
  /** @nullable */
  childName?: string | null;
  shippingAddress?: Address;
  /** @nullable */
  razorpayOrderId?: string | null;
  items: OrderItem[];
  createdAt: string;
}