# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TreasureTots Creations e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  /** @nullable */
  price?: number | null;
  stock?: number;
  coverImage: string;
  images: string[];
  category: string;
  /** @nullable */
  subcategory?: string | null;
  slug: string;
  isBuyable?: boolean;
}