# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TreasureTots Creations e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  role: string;
  createdAt?: string;
}